+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - core: Remove default.environments.yaml of modules an create a template for all modules.
 - core: Add support for multivhost.
 - core: Delete script for remove alpha jefe version.
+- core: Use DIR var for root dir of jefe-cli bash script
 # Version 1.3.1
 - module: Fix error do not up nginx to restart proyect for laravel and php-nginx-mysql module.
 - module: Fix do not working commands ps, restart, logs