+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - core: Use PROJECT_DIR constant for dir path of the jefe proyect
 - fix: Validate if www-data user exist when assigning permissions
 - docs: Add documentations of the task stop_nginx_proxy and start_nginx_proxy
+- core: Remove build task
 # Version 1.3.1
 - module: Fix error do not up nginx to restart proyect for laravel and php-nginx-mysql module.
 - module: Fix do not working commands ps, restart, logs